Changed All 5 pages to point to master file
I went and changed the references to point to the root.
Because this is my first time using GitHub I did NOT also include my
code in the home page yet, just in case i messed something up.  It does
not matter though because I am going to be working on it and need to
make changes tomorrow.  I will add it then.
Thanks.