Added my PerCapita HeatMap to Main Page
I lost alot of time getting this to actually load in the web page.  The
auto generated code is so bad to read, and there were so many div
objects, that I had to spend almost 2 hours exhaustively inserting my
code between different div statements.  If anyone else is having trouble
with this I can help.  Also, the part where I inserted my code is set to
1 pixel! so I almost didn't see that I was in the right place.  I can
help if you are having problems with this!